+++ catalyst/contrib/dl/callbacks/alchemy_logger.py
@@ -59,9 +59,9 @@ class AlchemyLogger(Callback):
             log_on_epoch_end (bool): logs per-epoch metrics if set True
         """
         super().__init__(
-            order=CallbackOrder.Logging,
-            node=CallbackNode.Master,
-            scope=CallbackScope.Experiment,
+            order=CallbackOrder.logging,
+            node=CallbackNode.master,
+            scope=CallbackScope.experiment,
         )
         self.metrics_to_log = metric_names
         self.log_on_batch_end = log_on_batch_end
@@ -105,9 +105,9 @@ class AlchemyLogger(Callback):
         """Translate batch metrics to Alchemy."""
         if self.log_on_batch_end:
             mode = runner.loader_name
-            metrics_ = runner.batch_metrics
+            metrics = runner.batch_metrics
             self._log_metrics(
-                metrics=metrics_,
+                metrics=metrics,
                 step=runner.global_sample_step,
                 mode=mode,
                 suffix=self.batch_log_suffix,
@@ -117,9 +117,9 @@ class AlchemyLogger(Callback):
         """Translate loader metrics to Alchemy."""
         if self.log_on_epoch_end:
             mode = runner.loader_name
-            metrics_ = runner.loader_metrics
+            metrics = runner.loader_metrics
             self._log_metrics(
-                metrics=metrics_,
+                metrics=metrics,
                 step=runner.global_epoch,
                 mode=mode,
                 suffix=self.epoch_log_suffix,