+++ catalyst/tools/meters/tests/test_ppv_tpr_f1.py
@@ -72,6 +72,10 @@ def runs_tests_on_meter_counts_and_value(meter, num_tp_check=16):
     """
     Tests the meter's counts and values (ppv, tpr, f1). Assumes there are no
     fp and fn (everything is tp).
+
+    Args:
+        meter:  meter instance
+        num_tp_check: number of tp checks
     """
     counts_dict = meter.tp_fp_fn_counts
     assert counts_dict["tp"] == num_tp_check