+++ CHANGELOG.md
@@ -8,11 +8,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- `log` parameter to `WandbLogger` ([#836](https://github.com/catalyst-team/catalyst/pull/836))
+- hparams experiment property ([#839](https://github.com/catalyst-team/catalyst/pull/839))
 - add docs build on push to master branch ([#844](https://github.com/catalyst-team/catalyst/pull/844))
 
 ### Changed
 
-- 
+- Renaming ([#837](https://github.com/catalyst-team/catalyst/pull/837))
+    - `SqueezeAndExcitation` -> `cSE`
+    - `ChannelSqueezeAndSpatialExcitation` -> `sSE`
+    - `ConcurrentSpatialAndChannelSqueezeAndChannelExcitation` -> `scSE`
+    - `_MetricCallback` -> `IMetricCallback`
+    - `dl.Experiment.process_loaders` -> `dl.Experiment._get_loaders`
+- `LRUpdater` become abstract class ([#837](https://github.com/catalyst-team/catalyst/pull/837))
+- `calculate_confusion_matrix_from_arrays` changed params order ([#837](https://github.com/catalyst-team/catalyst/pull/837))
 
 ### Removed